Prepare for relase using OpenStack release process
xstatic-spin has never been released using the OpenStack release
prociess with openstack/releases repository. To sync the release
history in openstack/releases and PyPI, we need to release a new
version, so the xstatic version is bumped.

README is also updated to match the length of the title marker
to the length of the title.

Change-Id: I27f5846730d2073bb39641b751461f410dab0a3c